+++ app/src/main/java/ru/alexbykov/revoluttest/currencies/domain/entity/CurrencyDetail.kt
@@ -3,6 +3,6 @@ package ru.alexbykov.revoluttest.currencies.domain.entity
 data class CurrencyDetail(
     val code: String,
     val displayName: String,
-    val exchangeRate: Float,
-    val calculatedValue: Float
+    val exchangeRate: Double,
+    val calculatedValue: Double
 )
